[rolling] image_publisher: Fix image, constantly flipping when static image is published (#986)
Continuation of
https://github.com/ros-perception/image_pipeline/pull/984.

When publishing video stream from a camera, the image was flipped
correctly. Yet for a static image, which was loaded once, the flip
function was applied every time `ImagePublisher::doWork()` was called,
resulting in the published image being flipped back and forth all the
time.

This PR should be straightforward to port it to `Humble`, `Iron` and
`Jazzy`.